# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Erlume Backend API",
    "version": "1.0.0",
    "description": "REST API for Erlume. Use this spec for documentation, code generation, or API clients. All IDs are MongoDB ObjectIds unless noted. List endpoints support date filters (year, month) where applicable."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "/",
      "description": "Relative (use your API base URL)"
    }
  ],
...
  "paths": {
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List users",
        "parameters": [
          {
            "name": "includeDeleted",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "description": "Include soft-deleted users"
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by role (e.g. seller)"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-usedBy": [
          "backoffice"
        ]
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create user",
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "x-usedBy": [
          "backoffice"
        ]
      }
    },
...
    "/api/sellers": {
      "get": {
        "tags": [
          "Sellers"
        ],
        "summary": "List sellers",
        "parameters": [
          {
            "name": "includeDeactivated",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "description": "Include deactivated sellers (default: false)"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-usedBy": [
          "backoffice"
        ]
      },
      "post": {
        "tags": [
          "Sellers"
        ],
        "summary": "Create seller (user + seller)",
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "x-usedBy": [
          "backoffice"
        ]
      }
    },
...
    "/api/items": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "List items",
        "parameters": [
          {
            "name": "dropId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemStatus",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "e.g. available, sold, pending"
          },
          {
            "name": "categoryId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sellerId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "User ID of seller"
          },
          {
            "name": "authenticationStatus",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "authentic",
                "not_authentic"
              ]
            }
          },
          {
            "name": "returnStatus",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by uploadedAt"
          },
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-usedBy": [
          "backoffice",
          "frontend"
        ]
      },
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "Create item",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "basePrice",
                  "condition",
                  "uploadedAt",
                  "saleRate",
                  "itemStatus",
                  "color",
                  "size",
                  "itemName",
                  "quantity",
                  "brandName",
                  "imageUrls",
                  "category_id",
                  "listingPrice"
                ],
                "properties": {
                  "basePrice": {
                    "type": "string",
                    "description": "Base price"
                  },
                  "condition": {
                    "type": "string",
                    "enum": [
                      "Like New",
                      "Gently Used",
                      "Worn"
                    ],
                    "description": "Item condition"
                  },
                  "uploadedAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "saleRate": {
                    "type": "string",
                    "description": "Sale/discount rate"
                  },
                  "itemStatus": {
                    "type": "string",
                    "enum": [
                      "pending",
                      "approved",
                      "available",
                      "sold",
                      "rejected"
                    ],
                    "description": "Item status"
                  },
                  "color": {
                    "type": "string"
                  },
                  "size": {
                    "type": "string"
                  },
                  "itemName": {
                    "type": "string",
                    "description": "Item name (or use bag)"
                  },
                  "bag": {
                    "type": "string",
                    "description": "Alias for itemName"
                  },
                  "itemModel": {
                    "type": "string"
                  },
                  "year": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "string"
                  },
                  "brandName": {
                    "type": "string",
                    "description": "Brand name (single field; do not send brand)"
                  },
                  "imageUrls": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Image URLs"
                  },
                  "photoUrls": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Alias for imageUrls"
                  },
                  "receiptPhotoUrls": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "priceEstimatorUrls": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "quoteUrls": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "approved": {
                    "type": "boolean"
                  },
                  "approvedNextDrop": {
                    "type": "boolean"
                  },
                  "authNeeded": {
                    "type": "boolean"
                  },
                  "cleaningNeeded": {
                    "type": "boolean"
                  },
                  "listingPrice": {
                    "type": "string",
                    "description": "Listing price (required); basePrice can differ"
                  },
                  "photographed": {
                    "type": "boolean"
                  },
                  "category_id": {
                    "type": "string",
                    "description": "Category ObjectId"
                  },
                  "sub_category_id": {
                    "type": "string"
                  },
                  "drop_id": {
                    "type": "string"
                  },
                  "sellerId": {
                    "type": "string",
                    "description": "Seller user ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "x-usedBy": [
          "backoffice",
          "frontend"
        ]
      }
    },
...
    "/api/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "List orders",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Order status filter"
          },
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-usedBy": [
          "backoffice",
          "frontend"
        ]
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Create order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "user_id",
                  "orderItems"
                ],
                "properties": {
                  "user_id": {
                    "type": "string"
                  },
                  "order_status": {
                    "type": "string"
                  },
                  "orderItems": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                      "type": "object",
                      "required": [
                        "item_id"
                      ],
                      "properties": {
                        "item_id": {
                          "type": "string"
                        },
                        "quantity": {
                          "type": "number",
                          "minimum": 1
                        },
                        "is_returned": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              },
              "example": {
                "user_id": "USER_OBJECT_ID",
                "order_status": "Pending",
                "orderItems": [
                  {
                    "item_id": "ITEM_OBJECT_ID_1",
                    "quantity": 1
                  },
                  {
                    "item_id": "ITEM_OBJECT_ID_2",
                    "quantity": 2
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "x-usedBy": [
          "backoffice",
          "frontend"
        ]
      }
    },
...
    "/api/orderitems": {
      "get": {
        "tags": [
          "OrderItems"
        ],
        "summary": "List order items",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-usedBy": [
          "backoffice"
        ]
      },
      "post": {
        "tags": [
          "OrderItems"
        ],
        "summary": "Create order item",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "order_id",
                  "item_id"
                ],
                "properties": {
                  "order_id": {
                    "type": "string"
                  },
                  "item_id": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "number",
                    "minimum": 1
                  },
                  "is_returned": {
                    "type": "boolean"
                  }
                }
              },
              "example": {
                "order_id": "ORDER_OBJECT_ID",
                "item_id": "ITEM_OBJECT_ID",
                "quantity": 1,
                "is_returned": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "x-usedBy": [
          "backoffice"
        ]
      }
    },
...
    "/api/transactions": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "List transactions",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-usedBy": [
          "backoffice"
        ]
      },
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Create transaction",
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "x-usedBy": [
          "backoffice"
        ]
      }
    },
...
    "/api/expenses": {
      "get": {
        "tags": [
          "Expenses"
        ],
        "summary": "List expenses",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 12
            }
          },
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-usedBy": [
          "backoffice"
        ]
      },
      "post": {
        "tags": [
          "Expenses"
        ],
        "summary": "Create expense",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "cost",
                  "month"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Required (e.g. September 2026 — Supplies)"
                  },
                  "cost": {
                    "type": "string",
                    "description": "Required (e.g. 6)"
                  },
                  "currency": {
                    "type": "string",
                    "description": "Optional, default KWD"
                  },
                  "employee_id": {
                    "type": "string",
                    "description": "Optional, ref User"
                  },
                  "notes": {
                    "type": "string",
                    "description": "Optional (e.g. 36 Packaging Bags)"
                  },
                  "type": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "ExpenseType values: supplies, subscriptions, services, subscriptions,services (compound). Use GET /api/enums/expenseType for full list."
                  },
                  "month": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Required (e.g. September 2026)"
                  },
                  "paidBy": {
                    "type": "string",
                    "description": "Optional"
                  },
                  "isRecurring": {
                    "type": "boolean",
                    "description": "Optional, default false"
                  },
                  "phase": {
                    "type": "string",
                    "description": "Optional (e.g. Pre-launch)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "x-usedBy": [
          "backoffice"
        ]
      }
    },
...
    "/api/employees": {
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "List employees",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-usedBy": [
          "backoffice"
        ]
      },
      "post": {
        "tags": [
          "Employees"
        ],
        "summary": "Create employee",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "photo": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "description": "e.g. full-time"
                  },
                  "salaryActual": {
                    "type": "string"
                  },
                  "salaryProjected": {
                    "type": "string"
                  },
                  "user_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "x-usedBy": [
          "backoffice"
        ]
      }
    },
...
    "/api/incomes": {
      "get": {
        "tags": [
          "Incomes"
        ],
        "summary": "List incomes. amount = Erlume commission (bank receipt). Filter by orderId, itemId, sellerId, year, month.",
        "parameters": [
          {
            "name": "orderId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sellerId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "User ID of seller"
          },
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-usedBy": [
          "backoffice"
        ]
      },
      "post": {
        "tags": [
          "Incomes"
        ],
        "summary": "Create income. amount = Erlume commission (what got to Erlume's bank, excludes seller payout). Supports month, prelaunch_bag, order/item/seller links.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount"
                ],
                "properties": {
                  "amount": {
                    "type": "string",
                    "description": "Erlume commission — what actually got to Erlume's bank account (excludes seller payout)"
                  },
                  "month": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Month for reporting"
                  },
                  "prelaunch_bag": {
                    "type": "string",
                    "description": "Item reference text (any product type) when item_id not linked. Format: Name-Brand-Year"
                  },
                  "order_id": {
                    "type": "string"
                  },
                  "order_item_id": {
                    "type": "string"
                  },
                  "item_id": {
                    "type": "string"
                  },
                  "seller_id": {
                    "type": "string",
                    "description": "Seller doc _id or User id (both accepted)"
                  },
                  "platform": {
                    "type": "string"
                  },
                  "income_type": {
                    "type": "string",
                    "default": "sale"
                  },
                  "received_at": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "x-usedBy": [
          "backoffice"
        ]
      }
    },
...
    "/api/sales": {
      "get": {
        "tags": [
          "Sales"
        ],
        "summary": "List sales",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "x-usedBy": [
          "backoffice"
        ]
      },
      "post": {
        "tags": [
          "Sales"
        ],
        "summary": "Create sale. amount = total listing price; erlumeCommission and sellerPayout are auto-calculated from Item.saleRate. Order flow: order_id+order_item_id. Prelaunch: amount/buyer/bag_record.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "order_id": {
                    "type": "string"
                  },
                  "order_item_id": {
                    "type": "string"
                  },
                  "item_id": {
                    "type": "string"
                  },
                  "transaction_id": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "string",
                    "description": "Total listing price (gross)"
                  },
                  "listingPrice": {
                    "type": "string"
                  },
                  "erlumeCommission": {
                    "type": "string",
                    "description": "Auto-calculated: amount * item.saleRate"
                  },
                  "sellerPayout": {
                    "type": "string",
                    "description": "Auto-calculated: amount * (1 - item.saleRate)"
                  },
                  "buyer": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  },
                  "sale_date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "bag_record": {
                    "type": "string"
                  },
                  "invoice_number": {
                    "type": "string"
                  },
                  "invoice_url": {
                    "type": "string"
                  },
                  "payment_evidence_url": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "x-usedBy": [
          "backoffice"
        ]
      }
    },
...
  }
}